+++ Remnants/Behaviours/BodyGrabbableObject.cs
@@ -16,6 +16,64 @@ namespace Remnants.Behaviours
             _bodySuitBehaviour = GetComponent<BodySuitBehaviour>();
             if (_bodySuitBehaviour == null)
                 mls.LogError("Did not found BodySuitBehaviour.");
+
+        }
+
+        private void Start()
+        {
+            //base.Start();
+            //for (int i = 0; i < propColliders.Length; i++)
+            //{
+            //    propColliders[i].includeLayers = -2621449;
+            //}
+            
+            propColliders = base.gameObject.GetComponentsInChildren<Collider>();
+ 
+
+            originalScale = base.transform.localScale;
+
+            if (itemProperties.isScrap && RoundManager.Instance.mapPropsContainer != null)
+            {
+                radarIcon = UnityEngine.Object.Instantiate(StartOfRound.Instance.itemRadarIconPrefab, RoundManager.Instance.mapPropsContainer.transform).transform;
+            }
+
+            if (!itemProperties.isScrap)
+            {
+                HoarderBugAI.grabbableObjectsInMap.Add(base.gameObject);
+            }
+
+            originalScale = base.transform.localScale;
+            if (itemProperties.itemSpawnsOnGround)
+            {
+                startFallingPosition = base.transform.position;
+                if (base.transform.parent != null)
+                {
+                    startFallingPosition = base.transform.parent.InverseTransformPoint(startFallingPosition);
+                }
+
+                FallToGround();
+            }
+            else
+            {
+                fallTime = 1f;
+                hasHitGround = true;
+                reachedFloorTarget = true;
+                targetFloorPosition = base.transform.localPosition;
+            }
+
+            MeshRenderer[] componentsInChildren = base.gameObject.GetComponentsInChildren<MeshRenderer>();
+            for (int j = 0; j < componentsInChildren.Length; j++)
+            {
+                componentsInChildren[j].renderingLayerMask = 1u;
+            }
+
+            SkinnedMeshRenderer[] componentsInChildren2 = base.gameObject.GetComponentsInChildren<SkinnedMeshRenderer>();
+            for (int k = 0; k < componentsInChildren2.Length; k++)
+            {
+                componentsInChildren2[k].renderingLayerMask = 1u;
+            }
+
+            EnablePhysics(true);
         }
         #endregion
 
@@ -23,6 +81,11 @@ namespace Remnants.Behaviours
         public override void EquipItem()
         {
             base.EquipItem();
+        }
+
+        public override void Update()
+        {
+            base.Update();
         }
 
         protected override void __initializeVariables()