+++ Remnants/Main.cs
@@ -16,7 +16,7 @@ namespace Remnants
         #region Variables
         private const string modGUID = "KawaiiBone.Remnants";
         private const string modName = "Remnants";
-        private const string modVersion = "1.3.4";
+        private const string modVersion = "1.3.5";
 
         public static Remnants Instance;
         private readonly Harmony _harmony = new Harmony(modGUID);